+++ src/router.js
@@ -13,7 +13,7 @@ export function useRoutes(
   } = {}
 ) {
   // path is the browser url location
-  let [path, setPath] = useState(getFormattedPath)
+  let [path, setPath] = useState(getFormattedPath(basePath))
 
   useLocationChange(setPath, { basePath, inheritBasePath: !basePath })
 
@@ -24,7 +24,7 @@ export function useRoutes(
     matchTrailingSlash
   })
   // No match should not return an empty Provider, just null
-  // console.log('router check', basePath, rootPath)
+  // console.log('router check', basePath, path, Object.keys(routes))
   if (!route || path === null) return null
   return (
     <BasePathContext.Provider value={basePath}>
@@ -59,6 +59,7 @@ function matchRoute(
     () => Object.keys(routes).map(createRouteMatcher),
     [hashRoutes(routes)]
   )
+  if (path === null) return null
   // Hacky method for find + map
   let routeMatch
   routeMatchers.find(([routePath, regex, groups]) => {